app/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
psycopg2-binary==2.9.9
sentence-transformers==3.2.1
requests==2.32.3
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0